pyproject.toml
[project]
name = "warehouse-system"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiosqlite>=0.22.0",
    "alembic>=1.17.2",
    "asyncpg>=0.31.0",
    "fastapi>=0.125.0",
    "flask>=3.1.2",
    "google-generativeai>=0.8.6",
    "greenlet>=3.3.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "sqlmodel>=0.0.27",
    "uvicorn>=0.38.0",
]
